Enabled SW path for Rx and optimization in Tx comp path
1. RX Data Path is through USB direct endpoint instead
of BAM pipe debug_quirks BIT(14) added to choose the path.
0 - Rx Data uses BAM2BAM path
1 - Rx Data uses Direct USB endpoint (Non-BAM2BAM)

2. Re-ordered packets sent to netif by default
since it affects the TCP Tput due to delay in
getting the ordered packets back from IPA, while
iperf runs from A5 and uses AMPDU pipe for flushing
the re-ordered packet. With this fix, always ordered
packets will go to n/w layer instead of AMPDU pipe

debug_quirks BIT(13)
0 - Send to AMPDU pipe
1 - Send to Netif

3. Tx complete error (-12)
Tx complete path for BAM2BAM pipe is optimized to not to use
worker queue to process the Tx complete event from IPA driver
This is done because, IPA send this in process context, and
no need to have one more worker thread in wlan to process
this. The TX_URB_CNT is increased to 64 from 32, since IPA
is slow in sending the data, and we need to buffer it.
This may not happen when the HBM park mode is enabled.
(ie sends 3 packets at a time instead of 1 packet)

4. Implemeted the code review comments

5. AMSDU de-aggr enabled in Host for the SW RX path
This setting is done through host interest area, based
on this FW decide to do the AMSDU aggr in FW or host.

6. Cookie was not released in copy_skbexpand fail case
in data_tx and dummy pkt send, and it is fixed.

7. Disconnect ipa was not called during teardown from
QCMAP, and it is fixed.

Change-Id: I4b293d9d2d3550c7541f1602e7d27a7c0fb6107d