ath6kl: Allow/Deny AP profile commit in SCC case
Currently ACS on AP vif not following the STA/AP
vif's connected channel though both the vifs are
configured to be operating in same band. And also
with SCC enabled case, MCC still happens when STA/AP
starts first and AP next on a diff channel. With
this changes in case of SCC if the host is requested to
start AP with acs option or particular channel before
commiting the profile, fix check if there is any other
vif already in connected state and uses it's channel to
start AP rather than acs or deny the AP start if the ch
requested is different than connected vif's channel in
the same band.

CRs-fixed: 458855, 445789

Change-Id: I84a5a1abd65e92741dcb36c08aeae9eafe75c7a1