ath6kl: ACS with HT Capability in 11N mode
HT capability needs the secondary channel to be selected
for the given primary channel. when ACS is enabled primary
channel will be selected only after the AP commit, which
conflicts to select secondary channel. This fix gives the
preference to ACS first and then selects HTCAP later,
provided that channel supports the given HT mode.

Change-Id: Ia7c4f0da88c41da2d361c800702f1c31d3a19fdc