ath6kl: Default handler for IDLE_CLOSE_TIMEOUT_EVENT.
When the firmware is operating in the WOW Suspend mode
the "Idle close timer" is started in the firmware
in case there are no stations connected to the AP's.
When the timer expires the host is woken up as part of
the remote resume process and the IDLE_CLOSE_TIMEOUT_EVENT
will be received by the host driver. Customers have the
option of handling this event the way they choose. One
option could be to bring down WLAN in order to save power.
The current default handler is an empty one which will print
a debug message.

Change-Id: I7fb6cccbfd67e1a6862f290acd081d0b76ce6130