ath6kl: Enable interface specific Tx flow control
Added support to control the network traffic based on the cookie
allocated to each virtual wlan interface. This is done to have
better load balance while running iperf with high bandwidth.
This algorithm will be active for network based traffic and
does normal flow control for the forwarded packets internally

Change-Id: I8d20e3f9a556622c918648d650f842b7b714c5d4