ath6kl: Fix for null-pointer exception at ath6kl_client_power_save
When hostapd gets teardown all station info will be removed
for the corresponding vif's (AP) station list. At the same time
there is a chance of getting client_power_save event from the
firmware which was received before the hostapd get teardown. In
the event handler we are getting the station conn info.
Ideally we should check the validity of the conn info
In the above mentioned scenerio there is a possiblity
that this conn info can be a null value, So we need to check
the value to avoid accessing the null pointer.

Change-Id: I694e3c6831555fe4b9b045ef44edc97b01d2eab9