ath6kl: Fix for the IPA deregistering error
In STA-AP mode, when STA tries to connect to the RefAP, host
receives STA Disconnect event with zero MAC address while not
able to connect to the RefAP. So this event should not be sent
to IPA for adding the rules. Fix is provided not to send IPA message
for MAC address zero.

Change-Id: I2bef1b07941380ee039e94edc2bc9b96f0fbd450
Crs-Fixed: 489034