ath6kl: Number of VIF count in sync with FW
Number of VIF count in AR structure specify the number of active interfaces.
This value will be zero for single interface and 1 for dual interface
This creates confusion, and not in sync with FW,  also the VIF max
count is set to 2 for dual interface. So the Num VIF count is modified
to have 1 for single interface and 2 for dual interface and so on.

Change-Id: Ie1c0a3a7660955e41ca08d9fef87ee4126bc83cd