ath6kl: Rx Flow control using RxQ Thold setting
Rx Data path flow is controlled by restricting the URB
submitted to HSIC/USB. The RxQ thold value is set to 256,
means, URB will not be submitted if the RxQ length is
greater than 256. By doing this, even the UL is pumped
with more than 150Mbps, the flow is contolled and
STA to A5 and STA to STA communication is better.

Earlier the RxQ Thold is set in the rx complete path, now the
Thold is set in the io_work_comp in the process context
as well. By doing this, host will have pending skbs limited
to 256 for processing. This value can be increased if
there is any impact on the UL Tput.

In addition to this, the cookie count have been increased
to 288 from 180. This is done to manage the traffic, and
not to drop packets while the data rate is high.
Worst-case, last 10% cookie count will be used for high
priority AC,and BE/BK data will be dropped.

Change-Id: Idc1211405fc0fb4b34c2643f3f8754eafb32c2aa