ath6kl: Tx complete processed in ath6kl WQ
Tx complete from IPA calls the driver function directly
instead of queuing in the ath6kl WQ. Due to this we have seen
watch dog bite in host while testing in MCC.

So now the tx complete is processed in ath6kl WQ context.
This is as per the original design, and there is no impact
on the DL Tput due to this change.

Change-Id: I64b2dd4dc02a6845c3fa6d540c9bac0fd3b54e68