ath6kl: Delete IPA header during AP stop
When ath6kl is unloaded, firmware sends disconnect event for AP
mode. Existing code checks for disconnect reason and then only
deletes the IPA header. But that resulted in not deleting the IPA
header. Modified the code to check for AP MAC address and if
disconnect event MAC matches with AP MAC address then delete the
IPA header.

Change-Id: If8152dec25822952dd362dea411057e534c0d4b4