ath6kl: Set proper TOS value in IPA HW
IP packet, TOS value is in 3 MSB bits, existing code sets the
values in 3 LSB bits. Now modified the code according to IPA
requirement where in ath6kl passes 3 bit mask and 3 bit TOS
value. This will be used in IPA hardware to classify the IP
packets and use the right pipe to send the packet to firmware.

Change-Id: Iafc6af9c9149994f1e65b0f0850178115737ae64