ath6kl: Cookie_used correction done in STA-AP mode
In STA-AP mode, E2E traffic was not running in security mode.
Analyzed and found, when there is more than one VIF, cookie used
is going negative since some of the data packets are sent in
control pipe, but during tx_complete it is accounted in VIF2, and
not in VIF0. Basically all the control packets are accounted in
VIF0. Due to this VIF1 cookie_used was going negative. So to address
this issue, extra tag added while sending the data packet in control
pipe, and the same is checked in tx_complete to account those packets
in VIF0. By doing this the cookie_used count is not going negative,
and it is matching correctly. Also check added while decrementing
the cookie_used value, so that it will not go to negative.
After the above fix, in STA-AP security mode, E2E traffic works fine.

Change-Id: I95bcdf801eb5db04cb1cc771ef059ddfb316c1fb