ath6kl: Fix a race in cancelling mcc_event_ctrl_timer
ath6kl_wmi_shutdown() is called but mcc_event_ctrl_timer
timer is not deleted/stopped. Because of this the timer
may get triggered and tries to send a WMI command to firmware,
But WMI context is already freed. Cancel the mcc event ctrl
timer just before ath6kl_wmi_shutdown to avoid this race condition.

Change-Id: Iab10d4c6179495ffece55ce0a2e9049492335916
Signed-off-by: Prasanth Bhatta <bhattap@codeaurora.org>