ath6kl: Removing Num_vif check in sscan_start func
After cretaing ath6kl core, 'ar' variables are initialized.
Num_vif variable was not initialized, and it is fixed. Also
in sscan_start() function , if the #of interface is more
than 1 then we are not allowing the command. This was done
earlier since FW does not support WOW in concurrency  mode.
Now FW supports the same so the check for num_vif > 1 return,
can be removed from the sscan_start().

After the above changes scan tested in STA-AP and AP-AP
mode in SBA9 build, and found working.

Change-Id: I89bb3fe46a0dbcaabb67e31a0eb4018ebc452865