ath6kl: Override sta_bh_override in AP-AP Auto SCC
In 2.4G automatic switching to SCC mode, found second AP
is not switching to the same channel of first AP, when
switching from STA-AP to AP-AP mode.

This is due to sta_bh_override check before committing the ap profile
in wmi.c file. Now additional check added and the sta_bh_override
is allowed only in STA-AP mode, and in AP-AP mode, only for the
first AP.

By this change 2.4G automatic channel switch to SCC works fine.

Also added check to support 11B, 11G as well in the orginal algm
for automatic switching to SCC in 2.4G.

Change-Id: I4b82fcfe5c9af4d6291bc1d12d536681fcd56697