ath6kl: Crash while sending scan result to nl80211
Crash observed in nl80211 work_queue() scheduling, while
updating the scan results from cfg80211_scan_complete_event.
Analysed the code, and found vif_.scan_req pointer is updated
in scan fcuntion before checking for SCAN in progress.
Due to this the scan results are sent for the cancelled request
of nl80211. Because of this, we have seen sometimes crash in
nl80211 work_queue scheduling.

This is fixed now, by assigning the scan request after checking
for SCAN in progress. Tested in MTP, and found working.
Need to do stress test in customer full system as well.

Change-Id: Ia3b7f45c94ce50b29c98c51794ac444a564c9fae