ath6kl: Send exact proto reason code for auth fail
Currently driver sends exact protocol code for connection
failure when previous state of station is connected. When
previous state of station is not connected, driver sends
generic reacon code instead of protocol reason code.

This restricts appplication not to reconfigure station to
SHARED-WEP when OPEN-WEP authentication fails.

The change has been done in a way that driver will send exact
proto reason code in case of OPEN-WEP or SHARED-WEP besides
station interface is not in connected state.

CRs-fixed: 529707

Change-Id: I6f37b85047075075dd07da57953e5d5e2c7b5641
Signed-off-by: Prashanth Bhatta <bhattap@codeaurora.org>