ath6kl: Sending HT regulatory info to hostapd
Currently while passing HT regulatory info to hospatd/
application we are not sending the information related to
RADAR, IBSS, PASSIVE SCAN. In some channels RADAR detection
should be enabled since we dont support RADAR detection
we should disable that channel for hostapd to use. Now
since we are not passing this information hostapd will
think that HT40 is allowed in this channel and it will try
to start AP in that channel. Now made changes to send all
the information properly.

Change-Id: Ie81cac4eba12b9dc60302fc70c2a085db3e39b49