ath6kl: Set phymode before profile commit
Setting of 11A phymode in 2nd interface when ACS in
11G mode is going on in 1st interface leads to target
assert due to channel structures getting overwritten.
So set the phymode in AP mode when the actual commit
happens to prevent target assert. The wrong check that
was preventing 2nd interface to start in 11A a mode is
also fixed.

CRs-fixed: 547285

Change-Id: I12d06e2b588ca9f54deede290107fc693a08f136