ath6kl: MCC flag disable issue during disconnect
MCC enable/disable flag is updated based on the AP/STA connect
or disconnect time in AP-AP, STA-AP modes. Based on this flag
SW_ROUTING is enabled/disabled in IPA. But during client disconnect
MCC flag got disabled which leads to disable SW_ROUTING to host in
IPA. This is fixed now, by setting the MCC flag to FALSE only in
STA disconnect or AP disconnect and not while client disconnects.

Change-Id: Ife4e362e17449bc03441312e75fe8d0e01676093
CRs-Fixed: 556049