ath6kl: ap is not bringing up in HT40 with LTE
When LTE commands comes and if the AP is in affected
channel AP will switch to non-affected channels, so
profile commit will be called. As part of profile commit
one disconnect event will come for AP, when disconnect for
AP comes it will restore HT capabilities with the default
values which is HT20 in 2.4G band. So when LTE channel
switch happens AP will not bring up in HT40 mode.
Now removed restoring the HT capabilities during
AP disconnect, because restoring HT capabilities is taken
care in ath6kl_stop_ap.

Change-Id: Ibad979d921dbb5542fd662e719c85c96cf35bae7
CRs-Fixed: 575035