ath6kl: LTE fix for AP-AP mode
Currently in ap_profile_commit we have apap_override check
which tells whether more than one AP is there or not, due
to this AP-AP mode for LTE Channel Avoidance is not working.
Now removed check for apap_override and force SCC in case
of LTE Channel Avoidance.

This fix is there in canary branch, but direct porting is
not possible so, we have separate change in hummingbird.
When we are porting the ACS for all the channels FR we can
revert this change.

Change-Id: Iddaed4d69b0b0cb77f2c5a8f6309b27b294c75cf
CRs-Fixed: 576281