ath6kl: Second AP not starting with ACS in both AP
When starting both the AP's with ACS, second AP we will put
in hold and in first will continue for profile_commit. Once
the connect event comes for first AP, second AP's profile
commit will continue with the first AP channel. While taking
first AP channel we are not checking whether both the AP's
are in same phy_mode, due to this we are giving bad channel
for profile commit. So second AP is not beaconing. Now made
changes to check for phy_mode while taking first AP's
channel for profile commit.

Change-Id: Iafdc16f3ca250180162bdfc2636a337215afb9fc
CRs-fixed: 604917