ath6kl: Configuring HT40 from hostapd
Currently if we give HT40+/- decision on choosing secondary
channel is taken care in firmware. So this will conflict
with the hostapd configuration and during initial OBSS scan.
Now made changes to send the secondary channel information
from driver to firmware so that hostapd configured secondary
channel only is selected.

Change-Id: Ife372d1ed9aef0f61947842f68936cbb0919b07a
CRs-fixed: 614752