ath6kl: Intial ping loss in WIFI client to RNDIS
When the device is configured with USB tethering and WLAN
AP mode, the multicast and broadcast packets are cloned and
shared between RNDIS client and ath6kl driver. When ath6kl
driver adds LLC header, skb data is modified as a result
ARP request is not resolved from RNDIS client. Added code
to unshare the skb if it is cloned.

Change-Id: I70559d197bf3b53224d11c88eea3bcdb8c15d401
CRs-fixed: 618858