ath6kl: Killing urbs during failure in usb probe
When creating the bam pipes we are submitting one dummy urb
for each bam pipe. If any failure happens during creating
any bampipe we are not killing the submitted urbs. So
during unloading of the module usb will try to clear the
submitted urbs and it will calll the callback registred by
us during submitting the dummy urbs. Since this callback
will be invalid at that time host will crash. Now made
changes to kill the urbs during failure.

Change-Id: Ia56a704a855e2674d507cbe5e3491f80cb680527
CRs-fixed: 634432