ath6kl: AP beaconing after killing hostapd in ACS
When hostapd is started with ACS and killed before the
connect event from firmware. At this time CONNECTED bit is
not be set, so disconnect command is not sent to firmware.
Due to this firmware is beaconing even after killing hostapd.
Now made changes to send disconnect command to firmware
always.

Change-Id: Iec96fbe05a1395f3dc944a354d68fcdae202e969
CRs-fixed: 661206, 647014, 656086