ath6kl: AP beaconing after killing hostapd in ACS
When hostapd is started with ACS and killed before the
connect event from firmware. At this time CONNECTED bit is
not be set, so disconnect command is not sent to firmware.
Due to this firmware is beaconing even after killing hostapd.
Now made changes to send disconnect command to firmware
always.

Change-Id: I9f50f27e4d45f5a1736f63f41bf85e83f02c61bb
CRs-fixed: 661206, 647014, 656086