ath6kl: fix for multicast aggregation
In STA mode multicast frames should not go for aggregation
reorder. but we had wrong checks due to that in AP mode
also multicast frames were not going through re-order path.
This was causing whole in re-ordering and ping was getting
delayed when there is any multicast frame in middle.

Change-Id: I10bed9db2acf3f8b6277eb6034d2b590f28558e0
CRs-fixed: 691677