ALX: Free packets on XMIT fail & trim RX packets
IPA sends packets to ALX for DL xmit.
In case of failure we need to free the skb
otherwise it results in a memory leak.

Min size for ethernet packets is 64.
Devices pad extra zeros for packets smaller than 64.
On RX packets trim the size and reduce additional padding.

Add some debug command to ethtool

Change-Id: I0173ff645795171ea646cf2d58794fc882136b56
CRs-fixed: 756007